apps: greybus-utils: Fix broken custom manifest builds
Commit 32671b67 (greybus-utils: add a way to associate a
device_id with a manifest) changed the format of the 'MANIFEST'
makefile variable used in 'apps/greybus-utils/Makefile'.  It now
expects the destination Device ID and a comma to be prefixed to
the manifest source file name.  The commit automatically prefixed
the Device ID and comma to non-custom manifest names but not to
custom manifest names.  A user can still specify the Device ID and
comma in the custom manifest name they enter but this is very obscure
if not broken behavior since it overrides the value specified in
'CONFIG_DST_DEVID'.

Make custom and non-custom manifest file names work similarly by
prefixing 'CONFIG_DST_DEVID' and a comma to custom manifest names
as well as non-custom ones.

Testing Done: Built with custom and non-custom manifests
Bug Number: Trivial
Review URL:
Reviewed by:

Signed-off-by: Mark Greer <mgreer@animalcreek.com>