tsb: pinshare: Use PinShare correctly on ES2
The ES2 architecture has a different PinShare layout than ES1.
ES2 adds several PinShare bits for I2C and I2S and forces
PinShare bits 2, 8, and 9 to be zero.

To handle these changes, ensure that only the relevant PinShare
macros are defined for each architecture.  When neither ES1 nor
ES2 are defined, generate a compilation error so the user is made
aware of the issue (this is better than wondering why things don't
work).  The I2C and I2S drivers are updated accordingly.

PinShare bits 2, 8, and 9 are not explicitly set to zero but they
default to zero and are never set so this should be safe.

Testing Done: Ran build_ara_image.sh, testd I2C & I2S on ES1, I2C on ES2
Bug Number: SW-954
Review URL:
Reviewed by:

Signed-off-by: Mark Greer <mgreer@animalcreek.com>